+++ pysatCDAAC/instruments/cosmic_gps.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
-"""
-Loads data from the COSMIC satellite.
+"""Loads data from the COSMIC satellite.
 
 The Constellation Observing System for Meteorology, Ionosphere, and Climate
 (COSMIC) is comprised of six satellites in LEO with GPS receivers. The
@@ -21,13 +20,13 @@ name
 tag
     Select profile type, or scintillation, using one of the following keys:
     {'ionprf': 'Ionospheric Profiles',
-    'wetprf': 'Atmospheric profiles with moisture',
-    'atmprf': 'Atmospheric profiles without moisture',
-    'eraprf': 'ERA-40 Interim reanalysis data',
-    'gfsprf': 'NCEP operational analysis data',
-    'ionphs': 'Ionospheric excess phase',
-    'podtec': 'Absolute Total Electron Content and auxiliary data',
-    'scnlv1': 'S4 scintillation index and auxiliary data'}
+     'wetprf': 'Atmospheric profiles with moisture',
+     'atmprf': 'Atmospheric profiles without moisture',
+     'eraprf': 'ERA-40 Interim reanalysis data',
+     'gfsprf': 'NCEP operational analysis data',
+     'ionphs': 'Ionospheric excess phase',
+     'podtec': 'Absolute Total Electron Content and auxiliary data',
+     'scnlv1': 'S4 scintillation index and auxiliary data'}
 inst_id
     None supported
 altitude_bin
@@ -51,17 +50,16 @@ Warnings
 """
 
 import datetime as dt
-import os
-import requests
-import tarfile
+import functools
 
-import numpy as np
 import netCDF4
+import numpy as np
 import pandas as pds
-import pysat
-from pysat import logger
-from pysat.utils import files as futils
 import xarray as xr
+
+import pysat
+from pysat.utils import files as futils
+from pysatCDAAC.instruments.methods import general as mm_cdaac
 
 
 # ----------------------------------------------------------------------------
@@ -69,12 +67,6 @@ import xarray as xr
 
 platform = 'cosmic'
 name = 'gps'
-
-l1_tags = ['ionPhs', 'podTec', 'scnLv1']
-lower_l1_tags = [tag.lower() for tag in l1_tags]
-
-l2_tags = ['ionPrf', 'wetPrf', 'atmPrf', 'eraPrf', 'gfsPrf']
-lower_l2_tags = [tag.lower() for tag in l2_tags]
 
 tags = {'ionprf': 'Ionospheric Profiles',
         'wetprf': 'Atmospheric profiles with moisture',
@@ -89,18 +81,29 @@ inst_ids = {'': list(tags.keys())}
 
 pandas_format = False
 
+# Translate tags to levels and camelcase names
+tag_translation = {'ionphs': {'level': 'level1b', 'substr': 'ionPhs'},
+                   'podtec': {'level': 'level1b', 'substr': 'podTec'},
+                   'scnlv1': {'level': 'level1b', 'substr': 'scnLv1'},
+                   'ionprf': {'level': 'level2', 'substr': 'ionPrf'},
+                   'wetprf': {'level': 'level2', 'substr': 'wetPrf'},
+                   'atmprf': {'level': 'level2', 'substr': 'atmPrf'},
+                   'eraprf': {'level': 'level2', 'substr': 'eraPrf'},
+                   'gfsprf': {'level': 'level2', 'substr': 'gfsPrf'}}
 # ----------------------------------------------------------------------------
 # Instrument test attributes
 
-_test_dates = {'': {}.fromkeys(list(tags.keys()), dt.datetime(2008, 1, 1))}
+_test_dates = {'': {}.fromkeys(list(tags.keys()), dt.datetime(2014, 5, 1))}
 
 # ----------------------------------------------------------------------------
 # Instrument methods
 
 
 def init(self):
-    """Initializes the Instrument object with instrument specific values.
+    """Initialize the Instrument object with instrument specific values.
 
+    Note
+    ----
     Runs once upon instantiation.
 
     """
@@ -115,7 +118,7 @@ def init(self):
                      'https://cdaac-www.cosmic.ucar.edu/cdaac/doc/cosmic.html'))
     self.acknowledgements = ack
     self.references = refs
-    logger.info(ack)
+    pysat.logger.info(ack)
 
     return
 
@@ -216,7 +219,7 @@ def list_files(tag=None, inst_id=None, data_path=None, format_str=None):
     """
 
     estr = 'Building a list of COSMIC files, which can possibly take time. '
-    logger.info('{:s}~1s per 100K files'.format(estr))
+    pysat.logger.info('{:s}~1s per 100K files'.format(estr))
 
     # Note that Files.from_os() could be used here except for the fact
     # that there are multiple COSMIC files per given time.
@@ -227,11 +230,11 @@ def list_files(tag=None, inst_id=None, data_path=None, format_str=None):
     # Overloading revision and cycle keyword below
     if format_str is None:
         # COSMIC file format string
-        if tag in lower_l2_tags or (tag == 'ionphs'):
+        if tag_translation[tag]['level'] == 'level2' or (tag == 'ionphs'):
             format_str = ''.join(('*/*/*_C{revision:03d}.{year:04d}.',
                                   '{day:03d}.{hour:02d}.{minute:02d}.',
                                   'G{cycle:02d}_{version:04d}.????_nc'))
-        elif tag in lower_l1_tags:
+        elif tag_translation[tag]['level'] == 'level1b':
             format_str = ''.join(('*/*/*_C{revision:03d}.{year:04d}.',
                                   '{day:03d}.{hour:02d}.{minute:02d}.',
                                   '????.G{cycle:02d}.??_{version:04d}.????_nc'))
@@ -365,7 +368,7 @@ def load(fnames, tag=None, inst_id=None, altitude_bin=None):
         # offset time could be obtained by parsing the filenames as is done
         # in list files however load isn't passed `format_str`, thus this
         # solution wouldn't work in all cases.
-        if tag not in lower_l1_tags or (tag == 'ionphs'):
+        if tag_translation[tag]['level'] == 'level2' or (tag == 'ionphs'):
             # Add 1E-5 seconds to time based upon occulting_inst_id and an
             # additional 1E-6 seconds added based upon cosmic ID.
             # Get cosmic satellite ID.
@@ -397,7 +400,7 @@ def load(fnames, tag=None, inst_id=None, altitude_bin=None):
                                                    uts=utsec.values)
 
         # Rename index to time.
-        if tag in lower_l1_tags:
+        if tag_translation[tag]['level'] == 'level1b':
             # scnlv1 files already have a 2D time variable, it is a conflict.
             output = output.rename(time='profile_time')
         output = output.rename(index='time')
@@ -686,99 +689,12 @@ def load_files(files, tag=None, inst_id=None, coords=None):
     return output
 
 
-def download(date_array, tag, inst_id, data_path=None,
-             user=None, password=None):
-    """Download COSMIC GPS data.
-
-    Parameters
-    ----------
-    date_array : array-like
-        list of datetimes to download data for. The sequence of dates need not
-        be contiguous.
-    tag : str
-        Tag identifier used for particular dataset. This input is provided by
-        pysat. (default='')
-    inst_id : str
-        Instrument ID string identifier used for particular dataset. This input
-        is provided by pysat. (default='')
-    data_path : str
-        Path to directory to download data to. (default=None)
-    user : str or NoneType
-        User string input used for download. Provided by user and passed via
-        pysat. If an account is required for downloads this routine here must
-        error if user not supplied. (default=None)
-    password : str or NoneType
-        Password for data download. (default=None)
-
-    Note
-    ----
-    This routine is invoked by pysat and is not intended for direct use by
-    the end user.
-
-    """
-
-    if tag in lower_l2_tags:
-        level_str = 'level2'
-        matches = [utag for ltag, utag in zip(lower_l2_tags, l2_tags)
-                   if tag == ltag]
-        sub_str = matches[0]
-    elif tag in lower_l1_tags:
-        level_str = 'level1b'
-        matches = [utag for ltag, utag in zip(lower_l1_tags, l1_tags)
-                   if tag == ltag]
-        sub_str = matches[0]
-
-    for date in date_array:
-        logger.info('Downloading COSMIC data for ' + date.strftime('%D'))
-        yr, doy = pysat.utils.time.getyrdoy(date)
-        yrdoystr = '{year:04d}/{doy:03d}'.format(year=yr, doy=doy)
-
-        # Try re-processed data (preferred).
-        # Construct path string for online file.
-        dwnld = ''.join(("https://data.cosmic.ucar.edu/gnss-ro/cosmic1",
-                         "/repro2013/", level_str, "/", yrdoystr, "/",
-                         sub_str, '_repro2013',
-                         '_{year:04d}_{doy:03d}.tar.gz'.format(year=yr,
-                                                               doy=doy)))
-        try:
-            # Make online connection.
-            with requests.get(dwnld) as req:
-                req.raise_for_status()
-        except requests.exceptions.HTTPError:
-            # If response is negative, try post-processed data
-            # Construct path string for online file
-            dwnld = ''.join(("https://data.cosmic.ucar.edu/gnss-ro/cosmic1",
-                             "/postProc/", level_str, "/", yrdoystr, "/",
-                             sub_str, '_postProc',
-                             '_{year:04d}_{doy:03d}.tar.gz'))
-            dwnld = dwnld.format(year=yr, doy=doy)
-            try:
-                # Make online connection
-                with requests.get(dwnld) as req:
-                    req.raise_for_status()
-            except requests.exceptions.HTTPError as err:
-                estr = ''.join((str(err), '\n', 'Data not found'))
-                logger.info(estr)
-
-        # Copy request info to tarball file with generated name in `fname`.
-        fname = os.path.join(data_path,
-                             ''.join(('cosmic_', sub_str,
-                                      '_{year:04d}.{doy:03d}.tar')))
-        fname = fname.format(year=yr, doy=doy)
-        with open(fname, "wb") as local_file:
-            local_file.write(req.content)
-            local_file.close()
-        try:
-            # Uncompress files and remove tarball
-            tar = tarfile.open(fname)
-            tar.extractall(path=os.path.join(data_path, yrdoystr))
-            tar.close()
-
-        except tarfile.ReadError:
-            # If file cannot be read as a tarfile, then data does not exist.
-            # Skip this day since there is nothing left to do.
-            pass
-        # tar file must be removed (even if download fails)
-        os.remove(fname)
-
-    return
+download_tags = \
+    {'': {tag:
+          {'remote_dir': ''.join(('gnss-ro/cosmic1/repro2013/',
+                                  tag_translation[tag]['level'],
+                                  '/{year:4d}/{day:03d}/')),
+           'tar_name': ''.join((tag_translation[tag]['substr'],
+                               '_repro2013_{year:4d}_{day:03d}.tar.gz')),
+           'backup': ['repro2013', 'postProc']} for tag in tags.keys()}}
+download = functools.partial(mm_cdaac.download, supported_tags=download_tags)