+++ pysatCDAAC/instruments/methods/__init__.py
@@ -1,0 +1,3 @@
+"""Methods for pysatCDAAC instruments."""
+
+from pysatCDAAC.instruments.methods import general  # noqa F401