+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.0.3] - ???
+## [0.x.x] - 2021-xx-xx
+* Updated `cosmic_gps` to support xarray Datasets
+* Added `download` to general methods
 * Update links in documentation
 
 ## [0.0.2] - 2021-06-18
@@ -11,7 +13,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Update `cosmic_gps` data directory structure
 * Updated parsing of `cosmic_gps` filenames to ensure only the latest
   version on disk is retained.
-- Updated `cosmic_gps` to support xarray Datasets
 
 ## [0.0.1] - 2020-08-13
 * initial port of existing routines from pysat